fix: rename Subitem1 to Subitem
